test(fetcher): split system tests into per-fetcher modules + address review
Distributes fetcher-specific tests from the monolithic `system/mod.rs` test
block into `#[cfg(test)] mod tests` blocks colocated with each fetcher,
matching AGENTS.md's "tests live next to the code" rule. Cross-cutting
tests (registration list, shape/sample contract sweep, helper formatters,
env detection) stay in `mod.rs`.

- New `system/test_helpers.rs` (cfg-gated) shares `ctx_text`,
  `ctx_with_shape`, `EnvGuard`, and `detect_terminal_with` across all
  per-fetcher test modules. Submodules import via
  `super::super::test_helpers::*`.

- Per-fetcher test placement:
  - `monitor_cpu::tests` — load defaults / text shape (2 tests)
  - `monitor_memory::tests` — ratio default / entries / text formatting (3 tests)
  - `monitor_uptime::tests` — text emission (1 test)
  - `monitor_load::tests` — text default / entries shape (2 tests)
  - `monitor_processes::tests` — count cap / textblock formatting (2 tests)
  - `monitor_disk::tests` — sample body / fetch defaults / bars (3 tests)
  - `monitor_host::tests` — rollup row count / textblock format (2 tests)
  - `monitor_battery::tests` — 14 tests covering text format, badge status,
    state mapping, AC fallbacks, panic-free coverage; uses
    `pub(crate)`-elevated internals (`BatterySnapshot`, `BatteryState`,
    `format_battery_text`, etc.) which are now reachable only from
    within `system_*` test modules.
  - `info_host::tests` — terminal kind default / arch / placeholder
    on unknown kind (3 tests)

Address CodeRabbit review on #227:

- Drop `system_text_all_kinds_helper_runs_in_regular_suite`: it called
  the `#[ignore]`-d `live_system_text_all_kinds` from a regular test,
  defeating the ignore boundary and reintroducing host-dependent
  behaviour into `cargo test`.

- `SystemMonitorBattery::read_snapshot` no longer panics on a poisoned
  manager mutex. `lock().unwrap_or_else(|e| e.into_inner())` matches
  the recovery pattern used elsewhere in the codebase and keeps battery
  reads working after any prior panic while holding the mutex.
