+++ src/fetcher/system/mod.rs
@@ -14,6 +14,9 @@ use super::{Fetcher, RealtimeFetcher};
 
 pub mod dmi;
 
+#[cfg(test)]
+pub(crate) mod test_helpers;
+
 mod info_bios;
 mod info_board;
 mod info_cpu;
@@ -469,101 +472,13 @@ fn format_bytes(bytes: u64) -> String {
 
 #[cfg(test)]
 mod tests {
-    use super::super::{FetchContext, Safety};
+    use super::super::Safety;
     use super::info_host::HostInfoOptions as SystemOptions;
-    use super::monitor_battery::{
-        BatteryOptions, BatterySnapshot, BatteryState, BatteryTextKind, battery_badge,
-        battery_entries, format_battery_text, format_percent, map_battery_state,
-        no_battery_payload,
-    };
+    use super::monitor_battery::{BatteryOptions, BatteryTextKind, format_percent};
     use super::monitor_processes::PROCESS_TOP_COUNT;
+    use super::test_helpers::{EnvGuard, ctx_text, detect_terminal_with};
     use super::*;
-    use crate::paths::TEST_ENV_LOCK;
-    use crate::payload::Status;
     use crate::render::Shape;
-    use std::sync::Mutex;
-    use std::time::Duration;
-
-    struct EnvGuard {
-        _lock: std::sync::MutexGuard<'static, ()>,
-        restore: Vec<(&'static str, Option<String>)>,
-    }
-
-    impl EnvGuard {
-        fn set(pairs: &[(&'static str, Option<&str>)]) -> Self {
-            let lock = TEST_ENV_LOCK.lock().unwrap_or_else(|e| e.into_inner());
-            let mut restore: Vec<(&'static str, Option<String>)> = Vec::new();
-            for (key, value) in pairs {
-                if !restore.iter().any(|(k, _)| k == key) {
-                    restore.push((*key, std::env::var(key).ok()));
-                }
-                match value {
-                    Some(value) => unsafe { std::env::set_var(key, value) },
-                    None => unsafe { std::env::remove_var(key) },
-                }
-            }
-            Self {
-                _lock: lock,
-                restore,
-            }
-        }
-    }
-
-    impl Drop for EnvGuard {
-        fn drop(&mut self) {
-            self.restore.iter().for_each(|(key, value)| match value {
-                Some(value) => unsafe { std::env::set_var(key, value) },
-                None => unsafe { std::env::remove_var(key) },
-            });
-        }
-    }
-
-    const TERMINAL_ENV_KEYS: &[&str] = &[
-        "WT_SESSION",
-        "GHOSTTY_RESOURCES_DIR",
-        "KITTY_WINDOW_ID",
-        "TERM",
-        "ALACRITTY_WINDOW_ID",
-        "ALACRITTY_LOG",
-        "WEZTERM_PANE",
-        "TERM_PROGRAM",
-    ];
-
-    fn ctx_with_shape(shape: Option<Shape>) -> FetchContext {
-        FetchContext {
-            widget_id: "w".into(),
-            timeout: Duration::from_secs(1),
-            shape,
-            ..Default::default()
-        }
-    }
-
-    fn ctx_text(options: Option<&str>) -> FetchContext {
-        let options = options.map(|s| toml::from_str::<toml::Value>(s).unwrap());
-        FetchContext {
-            widget_id: "w".into(),
-            timeout: Duration::from_secs(1),
-            shape: Some(Shape::Text),
-            options,
-            ..Default::default()
-        }
-    }
-
-    fn detect_terminal_with(overrides: &[(&'static str, &'static str)]) -> String {
-        let pairs: Vec<_> = TERMINAL_ENV_KEYS
-            .iter()
-            .map(|key| {
-                (
-                    *key,
-                    overrides.iter().find_map(|(override_key, value)| {
-                        (*override_key == *key).then_some(*value)
-                    }),
-                )
-            })
-            .collect();
-        let _guard = EnvGuard::set(&pairs);
-        detect_terminal()
-    }
 
     fn assert_realtime_contract(
         fetcher: &dyn RealtimeFetcher,
@@ -816,53 +731,6 @@ mod tests {
     }
 
     #[test]
-    fn disk_sample_body_matches_documented_shapes() {
-        assert!(matches!(
-            SystemMonitorDisk.sample_body(Shape::Ratio),
-            Some(Body::Ratio(ratio))
-                if ratio.label.as_deref() == Some("disk") && ratio.value == 0.58
-        ));
-        assert!(matches!(
-            SystemMonitorDisk.sample_body(Shape::Text),
-            Some(Body::Text(text)) if text.value == "58% of 400 GB"
-        ));
-        assert!(matches!(
-            SystemMonitorDisk.sample_body(Shape::Bars),
-            Some(Body::Bars(bars))
-                if bars.bars.len() == 3
-                    && bars.bars[0].label == "/"
-                    && bars.bars[2].value == 200
-        ));
-    }
-
-    #[test]
-    fn battery_sample_body_matches_documented_shapes() {
-        let fetcher = SystemMonitorBattery::default();
-
-        assert!(matches!(
-            fetcher.sample_body(Shape::Ratio),
-            Some(Body::Ratio(ratio))
-                if ratio.label.as_deref() == Some("battery") && ratio.value == 0.87
-        ));
-        assert!(matches!(
-            fetcher.sample_body(Shape::Text),
-            Some(Body::Text(text)) if text.value == "87% • Charging • 1h 23m"
-        ));
-        assert!(matches!(
-            fetcher.sample_body(Shape::Entries),
-            Some(Body::Entries(entries))
-                if entries.items.len() == 5
-                    && entries.items[0].key == "charge"
-                    && entries.items[4].value.as_deref() == Some("97%")
-        ));
-        assert!(matches!(
-            fetcher.sample_body(Shape::Badge),
-            Some(Body::Badge(badge))
-                if badge.status == Status::Ok && badge.label == "87% · Charging"
-        ));
-    }
-
-    #[test]
     fn sample_bodies_return_none_for_unsupported_shapes() {
         let unsupported = Shape::Image;
 
@@ -1044,107 +912,6 @@ mod tests {
         assert!(wrapped.status.is_none());
         assert!(wrapped.format.is_none());
         assert!(matches!(wrapped.body, Body::Text(TextData { value }) if value == "hello"));
-    }
-
-    #[test]
-    fn cpu_load_defaults_to_ratio() {
-        let p = SystemMonitorCpu::new().compute(&ctx_with_shape(None));
-        assert!(matches!(p.body, Body::Ratio(_)));
-    }
-
-    #[test]
-    fn cpu_load_emits_text_when_requested() {
-        let p = SystemMonitorCpu::new().compute(&ctx_with_shape(Some(Shape::Text)));
-        assert!(matches!(p.body, Body::Text(_)));
-    }
-
-    #[test]
-    fn memory_defaults_to_ratio() {
-        let p = SystemMonitorMemory::new().compute(&ctx_with_shape(None));
-        assert!(matches!(p.body, Body::Ratio(_)));
-        if let Body::Ratio(r) = p.body {
-            assert!((0.0..=1.0).contains(&r.value));
-        }
-    }
-
-    #[test]
-    fn memory_entries_shape_has_three_rows() {
-        let p = SystemMonitorMemory::new().compute(&ctx_with_shape(Some(Shape::Entries)));
-        assert!(matches!(p.body, Body::Entries(_)));
-        if let Body::Entries(e) = p.body {
-            let keys: Vec<_> = e.items.iter().map(|i| i.key.as_str()).collect();
-            assert_eq!(keys, ["used", "total", "free"]);
-        }
-    }
-
-    #[test]
-    fn memory_text_shape_formats_used_over_total() {
-        let p = SystemMonitorMemory::new().compute(&ctx_with_shape(Some(Shape::Text)));
-        assert!(matches!(p.body, Body::Text(_)));
-        if let Body::Text(text) = p.body {
-            assert!(text.value.contains(" / "));
-        }
-    }
-
-    #[test]
-    fn uptime_emits_text() {
-        let p = SystemMonitorUptime.compute(&ctx_with_shape(None));
-        assert!(matches!(p.body, Body::Text(_)));
-        if let Body::Text(t) = p.body {
-            assert!(!t.value.is_empty());
-        }
-    }
-
-    #[test]
-    fn load_average_defaults_to_text() {
-        let p = SystemMonitorLoad.compute(&ctx_with_shape(None));
-        assert!(matches!(p.body, Body::Text(_)));
-    }
-
-    #[test]
-    fn load_average_entries_shape_has_three_windows() {
-        let p = SystemMonitorLoad.compute(&ctx_with_shape(Some(Shape::Entries)));
-        assert!(matches!(p.body, Body::Entries(_)));
-        if let Body::Entries(e) = p.body {
-            let keys: Vec<_> = e.items.iter().map(|i| i.key.as_str()).collect();
-            assert_eq!(keys, ["1min", "5min", "15min"]);
-        }
-    }
-
-    #[test]
-    fn system_rollup_emits_six_rows() {
-        let p = SystemMonitorHost::new().compute(&ctx_with_shape(None));
-        assert!(matches!(p.body, Body::Entries(_)));
-        if let Body::Entries(e) = p.body {
-            assert_eq!(e.items.len(), 6);
-        }
-    }
-
-    #[test]
-    fn system_info_host_defaults_to_terminal_kind() {
-        let p = SystemInfoHost.compute(&ctx_text(None));
-        assert!(matches!(p.body, Body::Text(_)));
-        if let Body::Text(t) = p.body {
-            assert!(!t.value.is_empty());
-        }
-    }
-
-    #[test]
-    fn system_info_host_emits_arch_when_requested() {
-        let p = SystemInfoHost.compute(&ctx_text(Some("kind = \"arch\"")));
-        assert!(matches!(p.body, Body::Text(_)));
-        if let Body::Text(t) = p.body {
-            assert_eq!(t.value, std::env::consts::ARCH);
-        }
-    }
-
-    #[test]
-    fn system_info_host_rejects_unknown_kind_to_placeholder() {
-        let p = SystemInfoHost.compute(&ctx_text(Some("kind = \"bogus\"")));
-        assert!(matches!(p.body, Body::Text(_)));
-        if let Body::Text(t) = p.body {
-            assert!(t.value.starts_with("⚠"));
-        }
     }
 
     #[test]
@@ -1331,232 +1098,5 @@ mod tests {
     fn memory_totals_cache_reports_nonzero_memory() {
         let totals = cached_memory_totals();
         assert!(totals.memory > 0);
-    }
-
-    #[test]
-    fn system_text_all_kinds_helper_runs_in_regular_suite() {
-        live_system_text_all_kinds();
-    }
-
-    #[test]
-    fn system_text_block_shape_returns_key_value_strings() {
-        let p = SystemMonitorHost::new().compute(&ctx_with_shape(Some(Shape::TextBlock)));
-        assert!(matches!(p.body, Body::TextBlock(_)));
-        if let Body::TextBlock(l) = p.body {
-            assert_eq!(l.lines.len(), 6);
-            assert!(l.lines.iter().all(|s| s.contains(": ")));
-        }
-    }
-
-    #[test]
-    fn process_top_respects_count_cap() {
-        let p = SystemMonitorProcesses::new().compute(&ctx_with_shape(None));
-        assert!(matches!(p.body, Body::Entries(_)));
-        if let Body::Entries(e) = p.body {
-            assert!(e.items.len() <= PROCESS_TOP_COUNT);
-        }
-    }
-
-    #[test]
-    fn process_top_text_block_shape_formats_rows() {
-        let p = SystemMonitorProcesses::new().compute(&ctx_with_shape(Some(Shape::TextBlock)));
-        assert!(matches!(p.body, Body::TextBlock(_)));
-        if let Body::TextBlock(block) = p.body {
-            assert!(block.lines.len() <= PROCESS_TOP_COUNT);
-            assert!(block.lines.iter().all(|line| line.ends_with('%')));
-        }
-    }
-
-    fn snapshot(charge: f64, state: BatteryState, secs: Option<u64>) -> BatterySnapshot {
-        BatterySnapshot {
-            charge,
-            state,
-            time_remaining_secs: secs,
-            cycle_count: Some(284),
-            health: Some(0.97),
-        }
-    }
-
-    #[test]
-    fn battery_summary_includes_time_when_available() {
-        let snap = snapshot(0.87, BatteryState::Charging, Some(83 * 60));
-        assert_eq!(
-            format_battery_text(&snap, BatteryTextKind::Summary),
-            "87% • Charging • 1h 23m"
-        );
-    }
-
-    #[test]
-    fn battery_summary_omits_time_when_missing() {
-        let snap = snapshot(1.0, BatteryState::Full, None);
-        assert_eq!(
-            format_battery_text(&snap, BatteryTextKind::Summary),
-            "100% • Full"
-        );
-    }
-
-    #[test]
-    fn battery_text_kinds_pick_distinct_fields() {
-        let snap = snapshot(0.5, BatteryState::Discharging, Some(45 * 60));
-        assert_eq!(format_battery_text(&snap, BatteryTextKind::Percent), "50%");
-        assert_eq!(
-            format_battery_text(&snap, BatteryTextKind::Status),
-            "Discharging"
-        );
-        assert_eq!(
-            format_battery_text(&snap, BatteryTextKind::TimeRemaining),
-            "45m"
-        );
-    }
-
-    #[test]
-    fn battery_time_remaining_dash_when_missing() {
-        let snap = snapshot(1.0, BatteryState::Full, None);
-        assert_eq!(
-            format_battery_text(&snap, BatteryTextKind::TimeRemaining),
-            "—"
-        );
-    }
-
-    #[test]
-    fn battery_state_mapping_and_labels_cover_all_variants() {
-        use starship_battery::State;
-
-        assert_eq!(map_battery_state(State::Charging).label(), "Charging");
-        assert_eq!(map_battery_state(State::Discharging).label(), "Discharging");
-        assert_eq!(map_battery_state(State::Full).label(), "Full");
-        assert_eq!(map_battery_state(State::Empty).label(), "Empty");
-        assert_eq!(map_battery_state(State::Unknown).label(), "Unknown");
-    }
-
-    #[test]
-    fn battery_entries_include_optional_fields_only_when_present() {
-        let with = snapshot(0.5, BatteryState::Charging, Some(60));
-        let mut without = snapshot(0.5, BatteryState::Charging, None);
-        without.cycle_count = None;
-        without.health = None;
-        assert_eq!(battery_entries(&with).len(), 5);
-        assert_eq!(battery_entries(&without).len(), 2);
-    }
-
-    #[test]
-    fn no_battery_ratio_is_full_ac() {
-        let p = no_battery_payload(Shape::Ratio, BatteryTextKind::Summary);
-        assert!(matches!(p.body, Body::Ratio(_)));
-        if let Body::Ratio(r) = p.body {
-            assert_eq!(r.value, 1.0);
-            assert_eq!(r.label.as_deref(), Some("AC"));
-        }
-    }
-
-    #[test]
-    fn no_battery_text_varies_by_kind() {
-        let summary = no_battery_payload(Shape::Text, BatteryTextKind::Summary);
-        let percent = no_battery_payload(Shape::Text, BatteryTextKind::Percent);
-        let time = no_battery_payload(Shape::Text, BatteryTextKind::TimeRemaining);
-        assert!(matches!(summary.body, Body::Text(_)));
-        assert!(matches!(percent.body, Body::Text(_)));
-        assert!(matches!(time.body, Body::Text(_)));
-        if let Body::Text(text) = summary.body {
-            assert_eq!(text.value, "AC");
-        }
-        if let Body::Text(text) = percent.body {
-            assert_eq!(text.value, "100%");
-        }
-        if let Body::Text(text) = time.body {
-            assert_eq!(text.value, "—");
-        }
-    }
-
-    #[test]
-    fn no_battery_entries_and_badge_use_ac_placeholders() {
-        let entries = no_battery_payload(Shape::Entries, BatteryTextKind::Summary);
-        let badge = no_battery_payload(Shape::Badge, BatteryTextKind::Summary);
-        assert!(matches!(entries.body, Body::Entries(_)));
-        assert!(matches!(badge.body, Body::Badge(_)));
-        if let Body::Entries(entries) = entries.body {
-            assert_eq!(entries.items[0].key, "power");
-            assert_eq!(entries.items[0].value.as_deref(), Some("AC"));
-        }
-        if let Body::Badge(badge) = badge.body {
-            assert_eq!(badge.status, Status::Ok);
-            assert_eq!(badge.label, "AC");
-        }
-    }
-
-    #[test]
-    fn battery_compute_never_panics_on_any_shape() {
-        let f = SystemMonitorBattery::new();
-        for shape in [
-            None,
-            Some(Shape::Ratio),
-            Some(Shape::Text),
-            Some(Shape::Entries),
-            Some(Shape::Badge),
-        ] {
-            let p = f.compute(&ctx_with_shape(shape));
-            // Each branch must produce *some* body; on hosts without a battery we land on the
-            // AC stand-in, on laptops we get the real reading. Both are valid.
-            assert!(!matches!(p.body, Body::Image(_)));
-        }
-    }
-
-    #[test]
-    fn battery_badge_status_reflects_charge_and_state() {
-        let low = snapshot(0.05, BatteryState::Discharging, None);
-        assert_eq!(battery_badge(&low).status, Status::Error);
-        let mid = snapshot(0.30, BatteryState::Discharging, None);
-        assert_eq!(battery_badge(&mid).status, Status::Warn);
-        let high = snapshot(0.95, BatteryState::Discharging, None);
-        assert_eq!(battery_badge(&high).status, Status::Ok);
-        let charging_low = snapshot(0.05, BatteryState::Charging, None);
-        assert_eq!(battery_badge(&charging_low).status, Status::Ok);
-    }
-
-    #[test]
-    fn battery_rejects_unknown_option_to_placeholder() {
-        let f = SystemMonitorBattery::new();
-        let p = f.compute(&ctx_text(Some("bogus = true")));
-        assert!(matches!(p.body, Body::Text(_)));
-        if let Body::Text(t) = p.body {
-            assert!(t.value.starts_with("⚠"));
-        }
-    }
-
-    #[test]
-    fn battery_compute_without_manager_uses_no_battery_fallbacks() {
-        let fetcher = SystemMonitorBattery {
-            manager: Mutex::new(None),
-        };
-
-        let text = fetcher.compute(&ctx_text(Some("kind = \"percent\"")));
-        let entries = fetcher.compute(&ctx_with_shape(Some(Shape::Entries)));
-        let badge = fetcher.compute(&ctx_with_shape(Some(Shape::Badge)));
-        assert!(matches!(text.body, Body::Text(_)));
-        assert!(matches!(entries.body, Body::Entries(_)));
-        assert!(matches!(badge.body, Body::Badge(_)));
-        if let Body::Text(text) = text.body {
-            assert_eq!(text.value, "100%");
-        }
-        if let Body::Entries(entries) = entries.body {
-            assert_eq!(entries.items[0].value.as_deref(), Some("AC"));
-        }
-        if let Body::Badge(badge) = badge.body {
-            assert_eq!(badge.label, "AC");
-        }
-    }
-
-    #[tokio::test]
-    async fn disk_defaults_to_ratio_or_text_fallback() {
-        let ctx = ctx_with_shape(None);
-        let p = SystemMonitorDisk.fetch(&ctx).await.unwrap();
-        assert!(matches!(p.body, Body::Ratio(_) | Body::Text(_)));
-    }
-
-    #[tokio::test]
-    async fn disk_bars_shape_emits_bars_body() {
-        let ctx = ctx_with_shape(Some(Shape::Bars));
-        let p = SystemMonitorDisk.fetch(&ctx).await.unwrap();
-        assert!(matches!(p.body, Body::Bars(_)));
     }
 }
